Start organizing and cleaning up.
